Properly generate board in `create_full_board`
I simultaneously feel like the smartest and the dumbest programmer alive right now. When generating the board, I simply used list multiplication: `[[None] * 8] * 4`. Should make four rows of eight `None`s each, right? Well, it does, but it turns out every row is the same list in memory. This means that modifying one of those rows also modifies all the others.
(I feel like the smartest programmer alive for finding the cause of this, but also the dumbest for making it in the first place.)
Anyway, replace it with a list comprehension—what I should have done from the start. Generates a new list using the list literal.